Begin 3D lighting implementation/testing, use MSVC

+++ src/tiny_engine/framebuffer.cpp
@@ -48,8 +48,8 @@ void FullscreenFrameBuffer::DrawToScreen(const Shader& shader) {
     BindDefaultFrameBuffer();
     ClearGLColorBuffer();
     shader.use();
-    shader.setUniform("screenWidth", Camera::GetScreenWidth());
-    shader.setUniform("screenHeight", Camera::GetScreenHeight());
+    shader.setUniform("screenWidth", (f32)Camera::GetScreenWidth());
+    shader.setUniform("screenHeight", (f32)Camera::GetScreenHeight());
     shader.setUniform("time", (f32)GetTime());
     fullscreenSprite.DrawSprite(
         Camera::GetMainCamera(),
